stop persisting the work in EmbargoActor#destroy because the AccessControlList persists it for us.
something is still going on weird though. it appears that the embargo object in the database is different than the embargo metadata thats getting saved on the work. this means that the "past embargoes" section on the embargo/edit page shows different data than the data in the "deactivated embargoes" tab on the embargo/index page.

co-authored-by: braydon <braydon.justice@1268456bcltd.ca>