refactor flaky `valkyrie_ingest_job` spec
this test was being setup on the wrong unit (see removed comment), increasing
the complexity. it was running multiple layers of background jobs and relied on
a lot of configuration that wasn't very isolated.

the tested unit only needs to run against native valkyrie adapters (not Wings),
so it can be run as a unit test against the appropriate test adapters.