add a failing test for permission template grants on new objects
users in the PermissionTemplate for the adminset/workflow should get edit access
to new objects added to the admin set. test that at the controller level.

this was initially implemented as part of transactions in 05b66b90aaf, but
removed when shifting to Valkyrie because it wasn't tested anywhere. we couldn't
figure out what the desired behavior was, so we left it out of the new
implementation.

related to #5848.