add PermissionTemplate application to the `work_create` transaction
when we save a resource, we should run the `PermissionTemplateApplicator`
against it, to apply the template matching the resource's `AdministrativeSet`.

in the legacy Actor Stack, this step happens before the object is first
saved. in the Valkyrie model, we treat the ACL as a separate resource which can
(and has to) be saved separately, so it's convenient to first save the object
first, and then check for permission template application.

in the old model, we fail to save the work if a permission template doesn't
exist for the admin set. here, if we're missing an admin set or it is missing a
permission template, we simply decline to apply and always succeed (except in
cases of unhandled exceptions).