explicitly require `Retriable` for `UserStatImporter`
this whole class was failing silently because it raised `NameError` on
`Retriable` and then silently swallowed the exception.

this is only used in a Rake task, but since it gets loaded in production via
`app/**` anyway, we probably want to require its dependencies app wide. it seems
like this has been a real "problem" (does anyone use this rake task? its
function isn't totally clear to me) in all dockerized hyrax apps for some time(?).

this will need to be rewritten for Valkyrie, anyway, so getting a simple fix in
here seems appropriate.