actually release the embargo and lease with the embargo_lease rake tasks
with this commit, a user is able to run `rake hyrax:embargo:deactivate_expired` or `rake hyrax:lease:deactivate_expired` manually and only release the embargoes or leases whose expiration/release dates have already passed.
we are not simply calling the `deactivate!` method in each manager because that would deactivate ALL leases and embargoes, even if they are still active.

ref: #2657
ref: https://github.com/scientist-softserv/atla-hyku/issues/121