ensure both `EmbargoActor` and the embargo rake tasks set history
Hyrax counts on the `embargo_history` being present after an embargo is
released. move the logic for setting it to `EmbargoManager#release`. since
`#release` now involves updating both the `Hyrax::Embargo` and the
`Hyrax::AccessControl`, it's necessary to persist both after a release for it to
take effect. so far we've avoided having `EmbargoManager` handle persistence,
but maybe we want to change that to make this easier to use correctly at this
point?

----

this keeps the `nullify` call in `deactivate`, even though it should no longer
be needed. it's arguably a good idea if a user manually deletes an embargo to
empty the embargo's data---but then a little inconsistent to continue to use
that embargo to host the Work's embargo history(?).