never set `metadata/storage_adapter` to wings or AF if disable_wings
some test on `.koppie` seem to fail on test order(?) issues, when they decide to
use the Wings adapter for metadata persistence. it seems like something during
the test suite run is setting the adapter to `:wings_adapter` and then not
unsetting it. my best guess is that something is requiring `wings` or
`wings/setup` (maybe via `wings/helper`) during the test run.