test behavior after a published Fixity check failure on Valkyrie
if a Valkyrie Fixity service (to be invented) fails, it should be able to
publish an event to be picked up by `FileSetLifecycleNotificationListener` and a
notification can be published.

this updates the tests for the fixity failure notification path to avoid testing
AF behavior when Wings is disabled, and to add support for Valkyrie to the code
path.