only use specialized `Wings::WorkSearchBuilder` if Wings is enabled
`WorksControllerBehavior` offers the option to use it with
`curation_concern_type` that is a `Valkyrie::Resource`. in the event that one
does this, and `Wings` is turned on, it's necessary to provide a custom
search_builder_class to the controller to allow it to find resources indexed as
their mapped `ActiveFedora` objects.

that is, Valkyrie objects saved by Wings are indexed by ActiveFedora using model
names provided by AF.

none of this is needed if Valkyrie is handling the indexing, and the normal
search builder can be used in that case. so don't try to setup this class for
`.koppie` and in other cases where Wings is not in use.