Move force arg of collection_type_gid= to thread local var
PcdmCollection cannot handle the `force: true` argument supplied by the rake task, so move it to a thread local var for this special case. This is rather hacky, but avoids using a property and cluttering the data model. Using an instance variable to store the force state does not work in wings because it rebuilds the object from attributes when persisting.