Enable access of form classes, not just instances
Previously, there was no easy way to access the form for a given model
in the case where that model is not a PCDM object, as
`Hyrax::Forms::ResourceForm()` only worked in that case. This commit
moves the logic in `Hyrax::Forms::ResourceForm.for()` into
`Hyrax::Forms::ResourceForm()` so that the logic doesn’t need to be
duplicated in cases where just the class (and not a specific instance)
is desired.

It also makes `Hyrax::Forms::PcdmObjectForm()` callable and updates the
form generator to use that going forward. This simply makes the existing
behaviour of `Hyrax::Forms::ResourceForm()` more explicit; it already
produced a `Hyrax::Forms::PcdmObjectForm` subclass.