Move cleanup of work resource file sets from listener to transaction
If the list of member file sets is no longer available after a work resource is deleted, it will not be possible to perform the cleanup.
Therefore file sets (and their file metadata) must be deleted prior to the work's deletion during the transaction call.
The member cleanup listener is no long concerned with deleting file sets.