Clarify purpose of FileSetFileService
this service seems at first glance like it's just layering complexity on the
`#find_original_file` custom query (i almost deprecated it for this reason). but
its actually hiding an abstraction: the idea that the `pcdmuse:OriginalFile` is
the "primary" file for a FileSet.

this abstraction is useful, if Hyrax uses it consistently, because downstream
applications may want to determine the "primary" file any number of other
ways. (treating the OriginalFile as primary isn't at all in line with the PCDM
Use vocabulary's description of it, anyway).

clarify this all a bit with documentation and a method rename refactor.