Use class methods for Valkyrie model duck‐typing
It may be necessary to determine whether a model *class* represents a
PCDM Object, File Set, etc, not just a model *instance*. This commit
defines the former and aliases and deprecates the latter.

`collection?` and `pcdm_collection?` are split, in that
`Hyrax::AdministrativeSet` and `Hyrax::PcdmCollection` are both the
former, but not both the latter. Is this right, or should those
methods be synonyms?

`work?` is an alias for `pcdm_object? & !file_set?`.

`Hyrax::PcdmCollection` used to think it was a `pcdm_object?`, but it
shouldn’t have.

There’s probably a fair amount of `model.try(:work?)` which needs to be changed
to `model.class.try(:work?) || model.try(:work?)` before the old methods can be
removed. We’ll need to support both for as long as we support ActiveFedora, but
normalizing on a single place for querying this information, and having that
place be the model class, seemed desirable for Valkyrie resources.

Notably, this enables the use of these methods for determining which form or
indexer to use for a class of resource, instead of relying on class inheritance
checks.