Make AdministrativeSet a PCDM collection
It seems like there has been an attempt to consider administrative sets
as kinds of collections in the PCDM model, and we should push ahead with
that going forward.

This means that `:collection?` and `:pcdm_collection?` are effectively
synonyms. I prefer the latter (because it mirrors `:pcdm_object?`), but
it’s new and most Hyrax code currently uses the former.