Include basic metadata in CollectionResource
Instead of using a large list of attributes that *almost* match basic metadata, use the basic metadata schema as suggested in the generated collection metadata schema config.
The significant differences for koppie are creator becomes a required form field, and some fields are secondary instead of primary. Real applications are still free to customize their collection model as desired.