generalize Sipity::Entity() test for solr documents over classes
since Valkyrie doesn't support `#to_global_id`, we use a `ValkyrieGlobalIdProxy`
class for all `Valkyrie::ID` identified objects. hard coding the gid string to
include "GenericWork" in this test breaks for this reason when "GenericWork" is
a `Valkyrie::Resource`. to make it pass, we could build an actual work resource
and cast it using `Hyrax::GleobalID(resource)`, but since this test avoids real
models, we opt to just switch the expected entity match based on the model's
ancestry.