♻️ Leverage more of Blacklight's Solr URL
Prior to this commit, our defaults were not considering scheme, core,
password, nor user from the Blacklight URL.

What this meant was that the generated URL could strip away useful
information from the configured URL; in particular the user and password
information.

With this change, we utilize more of the underlying Blacklight
configuration; a net positive.