♻️ Extract Action Logic for improved composibility (#6644)
Prior to this commit, there was a precarity regarding the extensibility
of the create action for valkyrie objects.

The create method had several interacting objects, which made downstream
composibility harder to implement (without completely overriding the
action's behavior).  Downstream could:

1. Provide a different container for transactions
2. Inject additional steps into the transaction container, and those
   steps could require step args.
3. Remove/replace one of the transactions that has a step arg (which
   would then raise an error).

In all of these cases, the size of the original method makes for harder
downstream extensibility.

With this commit, I'm extracting an action object, putting forward a
minimum amount of work so we can discuss the approach.

This should be considered a noop change as I'm moving where the logic
happens.

If we deem that this is a good approach, I'd envision extracting some of
the work controller behavior specs into the given action.  I would also
anticipate extracting the update and delete actions.

Co-authored-by: Daniel Pierce <dlpierce@indiana.edu>