🐛 Fix Hyrax::ChangeSet.inspect method
Prior to this, there was no reference to `model_class` meaning that
`Hyrax::ChangeSet.inspect` would raise an exception.

We need the binding of the `resource_class`, which is available during
the creation of the class but not afterwards.  In other words, by using
the `define_singleton_method` we can leverage the provided
`resource_class` as a bound value for the return value of `.inspect`.