♻️ Favor Hyrax.query_service over method chaining
Prior to this commit we relied on the access_control object to "self
convert" via the `#valkyrie_resource` method.  However, that method
circumvents the `Hyrax.query_service` which provides a more up to date
version of the permissions.

This is subtle bug is most notable when we have composite adapters;
those modeled in the `double_combo` branch.

How it manifests is as follows:

- I have a Work and ACL in Fedora.
- I update the ACL to Postgres.
  - When I fetch the Work, it's in Fedora and grabs the ACL from Fedora
    which is not the most current.

This relates to:

- 0de7df6d2ee05f2d43b25acc3788349891a948d4
- https://github.com/samvera/hyrax/pull/6670