Lock down AdminSetCreateService class
The admin set create service cannot create an AdminSet. Changing to use
the config variable causes specs to break, because they are using the
default config value of AdminSet.

This is the first step to locking down the change. The default fallback
value to AdminSet should probably be changed as well.