🐛 internal_resource and to_rdf_representation are assumptive
Prior to this commit, when we attempted to query for all of the
collection models and admin set models, we'd see a mix of errors
depending on how we configured Wings; and regardless of configuring
Wings.

First, without a to_rdf_representation, we'd see an error like:

```
NoMethodError:
       undefined method `schema' for AdminSet:Class
```

Were we to configure Wings to handle an AdminSet, we'd then see

```
NoMethodError:
       undefined method `internal_resource' for #<AdminSet id: nil,
```

With this change, we're removing those errors.