🐛 Add conditional to `.fileset_for_directives`
We are noticing that when ingesting a PDF, the
`ValkyrieCreateDerivativesJob` is failing because the file_set doesn't
get found in `ValkyriePersistDerivatives.fileset_for_directives` when
the directives[:url] is an id instead of an actual uri, see
`FileSetDerivativesService#extract_full_text`.  I'm suggesting that we
check if we have a "/" and if not we assume that it is an id so we can
just look for that.