♻️ Favor source_type config check
With the usage of Hyrax.config.admin_set_class and collection_class, we
don't want to assume this previous hard-coded behavior.

To account for this change, we need to introduce some additional logic
to perform multiple searches.  While potentially non-performant,
hopefully the wings transformer logic can compress the models into a
single value.