🐛 Account for String and Valkyrie::ID Permissions
**tl;dr** Given that we sometimes might compare `object.id` for
permissions, we should treat Valkyrie::ID the same as the string.

Woe is us, for checking permissions on a String, what once was an
identifier, but which hath been usurped by the more rigid Valkyrie::ID.

But loe, for the old ways still work.  Granting permission to objects
based on a string identifier.  But Valkyrie::ID, one no longer a string,
fails.