🐛 Use configured class instead of hard-coded
The baseline configuration for the `Hyrax::PcdmMemberPresenterFactory`
allows for us to specify the `file_presenter_class` and the
`work_presenter_class`.  However, in the case of `presenter_for` we were
not leveraging that configurability.

This commit rectifies that behavior.

Tagging along with this commit is a minor performance improvement.
Namely, check a simple boolean before we begin checking elements that
might require additional queries and/or loading additional objects.