New query strategy to join FileSet solr documents
Fixes #6724
The previous attempt to join FileSet documents in the catalog search query failed to actually filter results. Due to a bad
or outdated test, this was not detected when running specs.

Updates the "term search" catalog controller spec to use the `all_fields` search. Prior to this change the spec did not follow
a code path to `join_for_works_from_files`. `all_fields` appears to be set for all queries originating from the hyrax web
interface (others are defined but unused in an unmodified CatalogController), so other specs that specify a `q` param
have had `all_fields` added.

Furthermore, the {!join} query did not have any effect when no FileSet documents are present in the solr index. Without a
FileSet in the mix, bad query results caused by the join were not detected by the spec.