♻️ Convert auto-file migration to job
Prior to this commit, we'd do this logic every time.  And part of that
logic (e.g. the actual moving of files) would be done during the request
cycle.

With this commit, we extract a job and the conditionals necessary to
avoid enqueuing lots of jobs that fundamentally would do nothing (e.g.
the files were already migrated)