♻️ Favor method_missing over a raft of delegates
It's like we're flooding a political convention with delegates.  Instead
of having all of those delegate methods, let's just implement method
missing and pass things along to the `SolrDocument`.

The benefit is that as we incorporate dynamic metadata, we don't need to
keep on adding `delegate :this_property, to: :solr_document`.