💄 Add custom cop for `include Hyrax::ArResource`
This commit will add a custom cop to RuboCop that will check for
`include Hyrax::ArResrouce` or `include ArRdResource` in the codebase.
The `ArResource` module is just for development ease of use and should
not be used within the codebase.

The custom cop leaves an example for additional custom cops that can be
used in the future and where to test them.  The key is to make sure the
`.rubocop.yml` file is updated to include the new cop.  The pattern for
it is adding something like Hyrax::RuboCop::CustomCops::SomeCop would be

```yaml
Hyrax/SomeCop:
  Enabled: true
```

From what I gather it is taking the top most namespace and then the
class name.