Get derivatives downloads working
This commit will tighten up the logic for a file set's derivatives to be
downloadable by the user by adding a #extensions_and_mime_types method
to the FileSet model and indexing it to the Solr document so the
presenter has access to it.  Though the method is named
"extensions_and_mime_types", we also give it some additional information
such as the use, id, and name.  Currently the use and id are not used
but might be useful in the future.