Address valkyrie derivative issues (#6795)
* Get derivatives downloads working

This commit will tighten up the logic for a file set's derivatives to be
downloadable by the user by adding a #extensions_and_mime_types method
to the FileSet model and indexing it to the Solr document so the
presenter has access to it.  Though the method is named
"extensions_and_mime_types", we also give it some additional information
such as the use, id, and name.  Currently the use and id are not used
but might be useful in the future.

* Address rubocop violation

* Make method handle missing data

Method `extensions_and_mime_types` failed when there was no
original_filename, making specs touchy. This PR lets it return cleanly.

* Final cleanup

- remove stray comment
- return array instead of hash on file_set

* Address PR feedback

- add specs for new method
- update solr index field suffic
- use download_url

* Add to_json in indexer

Dropped by accident...

---------

Co-authored-by: Kirk Wang <kirk.wang@scientist.com>