WIP to fix migrating files
We have to migrate original file before derivatives, so this switches
the order of migration.

We have to remove the id of the pre-migrated file_set's original file
before saving the new FileMetadata object's file_id. This prevents
creating duplicate ORM resources.

Indexing the file_set for the original file's FileMetadata object still
fails while trying to index the thumbnail.