Bring back hamburger menu on the top right
This commit will remove a button that appears on the facets at a certain
width that, as far as I can tell, doesn't seem to do anything at the
moment.  I think it's supposed to open and collapse the all the facets.
Consider bringing it back if the functionality is needed/fixed.