Expand support for i18n labels
This commit will add a few different conditions for populating the
label.  If the label is a string we'll simply set that as the label.

ex. `label: 'My Label'`

If the label is a hash we'll look for the key that matches the current
locale.

ex. `label: { en: 'My Label', es: 'Mi Etiqueta' }`

If the locale is not found it will fall back to en as the default, but
if en isn't found it will just use the first value in the hash.

ex. `label: { es: 'Mi Etiqueta', fr: 'Mon Étiquette' }`