Fix Add Child link (#6837)
* Fix Add Child link

The links under the Add Child dropdown was not working because the HTML
structure was not correct.  This commit will close the button tag and
then open the ul tag which allows the links to be clickable.

* Change the structure of work show, add child link

Since the bootstrap upgrade the structure of ul > il > a doesn't seem to
work like before.  The <a> tag was only clickable when you hover over
the text.  This change will make the whole <li> clickable much like the
nav dropdown menu for the My Profile/Dashboard/Logout links on the top
right of the masthead.