Some frontend updates (#6847)
* Make pagination adjust at an earlier width

The pagination links on the catalog results page were getting really
squished so this commit will make it move up above the sorting controls
at an earlier width.

* Bring back hamburger menu on the top right

This commit will remove a button that appears on the facets at a certain
width that, as far as I can tell, doesn't seem to do anything at the
moment.  I think it's supposed to open and collapse the all the facets.
Consider bringing it back if the functionality is needed/fixed.

* Make pagination colors same as primary color

The pagination colors were still default and doesn't match anything.
This commit will couple it with the primary button color.

* Break long links in catalog search

This will make any long links break.  Prior to this commit any links
that didn't have a `-` or a character that would create a natural word
break would push the UI until it looked weird.

* Make new/edit works form labels bold

In the new/edit works form, the labels used to be bold, but they are not
anymore. This commit will make it bold again.