Fix catalog search dropdown (#6867)
* Fix catalog search form dropdown not changing

On the home page when the user is logged in, the dropdown next to the
search was not changing when the user selects a different option.  This
was because the javascript was changing the data-search-element="label"
span but what is actually being shown is the aria-hidden="true" span.

Before:
  <span class="sr-only" data-search-element="label">All</span>
  <span aria-hidden="true">All</span>

  *user selects a different option*

  <span class="sr-only" data-search-element="label">My Works</span>
  <span aria-hidden="true">All</span>

After:
  <span class="sr-only" data-search-element="label">All</span>
  <span aria-hidden="true">All</span>

  *user selects a different option*

  <span class="sr-only" data-search-element="label">My Works</span>
  <span aria-hidden="true">My Works</span>

* Change html structure for search dropdown

This change will align how Bootstrap 4 does dropdown menus.  The
exmaples that I see in the documentation show that it should be in a
<div> tag with <a> tags inside of it.