Maps location for wings migrations
- Native Valkyrie objects contain a URI string, not an RDF::URI
- Fedora contains Hyrax::ControlledVocabularies::Location objects

Modifies the behavior of the `based_near` attribute in the `to_resource`
and `from_resource` methods. When using wings to convert between
resources & fedora works, we were previously storing the RDF::URI in
Valkyrie. This PR updates the behavior to consistently return values
consistent with what would be seen in native objects.