Fixes Wings migration of based_near attribute in to_resource (#6922)
* Add failing specs to fix

* Maps location for wings migrations

- Native Valkyrie objects contain a URI string, not an RDF::URI
- Fedora contains Hyrax::ControlledVocabularies::Location objects

Modifies the behavior of the `based_near` attribute in the `to_resource`
and `from_resource` methods. When using wings to convert between
resources & fedora works, we were previously storing the RDF::URI in
Valkyrie. This PR updates the behavior to consistently return values
consistent with what would be seen in native objects.

* Pacify the cops

* Undo Active Fedora conversion of based_near

This may still be needed, but this doesn't work correctly.
A ticket has been created for this bug if this is something that is
ever needed.

https://github.com/samvera/hyrax/issues/6926

---------

Co-authored-by: Rob Kaufman <rob@notch8.com>