Add fixes for GA4
In the base, we are adding a few nil guards to we don't end up with a
NoMethodError for NilClass. In the GA4 module we are adding a case where
if we are getting a Hash already, don't try to JSON.parse it.  In the
UserStatImporter we are switching a query to the Valkyrie way, which
should already be in main.